Refactor execution logic into shared helper
